+++ common/src/main/java/de/mrjulsen/wires/decoration/WireDecorationRenderer.java
@@ -1,6 +1,6 @@
 package de.mrjulsen.wires.decoration;
 
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
@@ -13,5 +13,5 @@ public abstract class WireDecorationRenderer<T extends IWireDecoration<T>> {
         this.decoration = decoratin;
     }
 
-    public abstract void render(PoseStack poseStack, VertexConsumer consumer, Vector3f pos, Vector3f direction, int light);
+    public abstract void render(PoseStack poseStack, VertexConsumer consumer, Vector3d pos, Vector3d direction, int light);
 }
